requirements.txt
unicorn==1.0.1
pyelftools==0.24
hexdump==3.3
keystone-engine==0.9.1.post3
capstone=4.0.1
frida=12.2.27